Refactor `structural*` files.
