+++ trie/resolve_set.go
@@ -21,18 +21,6 @@ import (
 	"sort"
 )
 
-type sortable [][]byte
-
-func (s sortable) Len() int {
-	return len(s)
-}
-func (s sortable) Less(i, j int) bool {
-	return bytes.Compare(s[i], s[j]) < 0
-}
-func (s sortable) Swap(i, j int) {
-	s[i], s[j] = s[j], s[i]
-}
-
 // ResolveSet encapsulates the set of keys that are required to be fully available, or resolved
 // (by using `BRANCH` opcode instead of `HASHER`) after processing of the sequence of key-value
 // pairs
@@ -95,28 +83,4 @@ func (rs *ResolveSet) HashOnly(prefix []byte) bool {
 		return false
 	}
 	return true
-}
-
-func generateStructLen(buffer []byte, l int) int {
-	if l < 56 {
-		buffer[0] = byte(192 + l)
-		return 1
-	}
-	if l < 256 {
-		// l can be encoded as 1 byte
-		buffer[1] = byte(l)
-		buffer[0] = byte(247 + 1)
-		return 2
-	}
-	if l < 65536 {
-		buffer[2] = byte(l & 255)
-		buffer[1] = byte(l >> 8)
-		buffer[0] = byte(247 + 2)
-		return 3
-	}
-	buffer[3] = byte(l & 255)
-	buffer[2] = byte((l >> 8) & 255)
-	buffer[1] = byte(l >> 16)
-	buffer[0] = byte(247 + 3)
-	return 4
 }
