+++ trie/rlphacks/bytes_test.go
@@ -1,4 +1,4 @@
-package trie
+package rlphacks
 
 import (
 	"bytes"
